TOFIXUP: DRY out shared controller code in concern
I tried this initially in a helper, however instance variables displease
rubocop there, which makes sense. Attempts to refactor these hit the
need for assignments to be made in before actions. In the end a concern
seemed like the easiest way to keep the code pretty much as is but
reduce duplication
