+++ app/controllers/account_subscriptions_controller.rb
@@ -1,6 +1,8 @@
 class AccountSubscriptionsController < ApplicationController
   include GovukPersonalisation::ControllerConcern
   include AccountHelper
+  include ContentItemNotifications
+
   before_action :enforce_feature_flag
   before_action :assign_content_item
   before_action :assign_list_params
@@ -37,23 +39,5 @@ private
 
   def enforce_feature_flag
     head :not_found and return unless govuk_account_auth_enabled?
-  end
-
-  def assign_content_item
-    # NOTE: the "topic" param has historically appeared in external links
-    @content_item_path = params[:topic] || params[:link]
-
-    return bad_request unless @content_item_path.to_s.starts_with?("/")
-    return bad_request unless URI.parse(@content_item_path).relative?
-
-    @content_item ||= GdsApi.content_store.content_item(@content_item_path)
-  rescue URI::InvalidURIError
-    bad_request
-  end
-
-  def assign_list_params
-    @list_params = GenerateSubscriberListParamsService.call(@content_item.to_h)
-  rescue GenerateSubscriberListParamsService::UnsupportedContentItemError
-    bad_request
   end
 end
